chore: Hide sidebar navigation on /edit-profile view
