+++ src/app/(proper_react)/(redesign)/(authenticated)/user/(dashboard)/settings/panels/SettingsPanelEditInfoRedesign.tsx
@@ -216,7 +216,7 @@ function SettingsPanelEditInfoRedesign(
     <>
       <div className={styles.header}>
         <div>
-          <h3>{l10n.getString("settings-tab-label-edit-profile")}</h3>
+          <h3>{l10n.getString("settings-tab-label-update-scan-info")}</h3>
           {props.profileData && (
             <p>
               {l10n.getFragment("settings-update-scan-info-description", {
