+++ src/app/(proper_react)/(redesign)/(authenticated)/user/(dashboard)/settings/stories/SettingsStoryWrapper.tsx
@@ -72,6 +72,7 @@ export const SettingsWrapper = (props: SettingsWrapperProps) => {
           countryCode={props.countryCode}
           enabledFeatureFlags={props.enabledFeatureFlags}
           experimentData={defaultExperimentData["Features"]}
+          hideSidebar={props.activeTab === "edit-profile"}
         >
           <SettingsView
             l10n={getL10n()}
