+++ src/app/(proper_react)/(redesign)/Shell/ShellRedesign.tsx
@@ -7,21 +7,14 @@ import Link from "next/link";
 import Image from "next/image";
 import { Session } from "next-auth";
 import { MobileShell } from "../MobileShell";
-import { PageLink } from "../PageLink";
 import { Footer } from "../Footer";
 import { ExtendedReactLocalization } from "../../../functions/l10n";
 import { FeatureFlagName } from "../../../../db/tables/featureFlags";
 import { ExperimentData } from "../../../../telemetry/generated/nimbus/experiments";
-import {
-  DashboardIcon,
-  NotificationIcon,
-  FaqIcon,
-  ScanInfoIcon,
-  SettingsIcon,
-  TipIcon,
-} from "../../../components/server/Icons";
 import MonitorLogo from "../../images/monitor-logo.svg";
 import styles from "./ShellRedesign.module.scss";
+import { NavbarList } from "./ShellNavbarList";
+import { UserAnnouncementsRow } from "knex/types/tables";
 
 export type Props = {
   l10n: ExtendedReactLocalization;
@@ -37,109 +30,9 @@ export type Props = {
     yearly: number;
     monthly: number;
   };
+  announcements: UserAnnouncementsRow[] | null;
+  hideSidebar?: boolean;
 };
-
-export const NavbarList = (props: {
-  l10n: ExtendedReactLocalization;
-  countryCode: string;
-  enabledFeatureFlags: FeatureFlagName[];
-}) => (
-  <div className={styles.navbarList}>
-    <ul className="noList">
-      <li key="dashboard">
-        <PageLink
-          href="/user/dashboard"
-          activeClassName={styles.isActive}
-          hasTelemetry={{ link_id: "navigation_dashboard" }}
-        >
-          <DashboardIcon alt="" />
-          {props.l10n.getString("main-nav-link-dashboard-label")}
-        </PageLink>
-      </li>
-      {props.enabledFeatureFlags.includes("EditScanProfileDetails") && (
-        <li key="settings-edit-info">
-          <PageLink
-            href="/user/settings/edit-info"
-            activeClassName={styles.isActive}
-            hasTelemetry={{ link_id: "navigation_edit_info" }}
-          >
-            <ScanInfoIcon alt="" />
-            {props.l10n.getString("settings-tab-label-edit-profile")}
-          </PageLink>
-        </li>
-      )}
-    </ul>
-    <hr />
-    <strong>{props.l10n.getString("main-nav-link-settings-label")}</strong>
-    <ul className="noList">
-      {!props.enabledFeatureFlags.includes("EditScanProfileDetails") && (
-        <li key="settings-edit-info">
-          <PageLink
-            href="/user/settings/edit-info"
-            activeClassName={styles.isActive}
-            hasTelemetry={{ link_id: "navigation_edit_info" }}
-          >
-            <ScanInfoIcon alt="" />
-            {props.l10n.getString("settings-tab-label-edit-info")}
-          </PageLink>
-        </li>
-      )}
-      <li key="settings-notifications">
-        <PageLink
-          href="/user/settings/announcements"
-          activeClassName={styles.isActive}
-          hasTelemetry={{
-            link_id: "navigation_settings_notifications",
-          }}
-        >
-          <NotificationIcon alt="" />
-          {props.l10n.getString("settings-tab-label-notifications")}
-        </PageLink>
-      </li>
-      <li key="settings-manage-account">
-        <PageLink
-          href="/user/settings/manage-account"
-          activeClassName={styles.isActive}
-          hasTelemetry={{
-            link_id: "navigation_settings_manage_account",
-          }}
-        >
-          <SettingsIcon alt="" />
-          {props.l10n.getString("settings-tab-label-manage-account")}
-        </PageLink>
-      </li>
-    </ul>
-    <hr />
-    <ul className="noList">
-      {props.countryCode === "us" && (
-        <li key="how-it-works">
-          <PageLink
-            href="/how-it-works"
-            activeClassName={styles.isActive}
-            target="_blank"
-            hasTelemetry={{ link_id: "navigation_how_it_works" }}
-          >
-            <TipIcon alt="" />
-            {props.l10n.getString(
-              "landing-premium-hero-navbar-link-how-it-works",
-            )}
-          </PageLink>
-        </li>
-      )}
-      <li key="faq">
-        <PageLink
-          href="https://support.mozilla.org/kb/firefox-monitor-faq"
-          title={props.l10n.getString("main-nav-link-faq-tooltip")}
-          target="_blank"
-          hasTelemetry={{ link_id: "navigation_faq" }}
-        >
-          <FaqIcon alt="" />
-          {props.l10n.getString("main-nav-link-faq-label")}
-        </PageLink>
-      </li>
-    </ul>
-  </div>
-);
 
 export const ShellRedesign = (props: Props) => {
   return (
@@ -166,11 +59,12 @@ export const ShellRedesign = (props: Props) => {
                 width={170}
               />
             </Link>
-            <NavbarList
-              l10n={props.l10n}
-              countryCode={props.countryCode}
-              enabledFeatureFlags={props.enabledFeatureFlags}
-            />
+            {!props.hideSidebar && (
+              <NavbarList
+                countryCode={props.countryCode}
+                enabledFeatureFlags={props.enabledFeatureFlags}
+              />
+            )}
           </nav>
           <div className={styles.page}>{props.children}</div>
         </div>
